Do some ad-hoc CSE in Exitification
So that code like this

    safe :: Int -> [Int] -> [[Int]]
    safe x2 ds6
      = case ds6 of
          [] -> lvl6;
          q : l | x2     == q -> h x2 ds5
                | x2 + 1 == q -> h x2 ds5
                | x2 + 2 == q -> h x2 ds5
                | x2 + 3 == q -> h x2 ds5
                | otherwise   -> safe (x+10) l

only gets one exit join point, instead of 4 identically.

(There is currently no CSE for join points, and even if there were, it
is cheap to do it here.)

This does happen at least in GHC somewhere (exhibited by a core lint
error when this patch was buggy). Nofib does not show any differences,
not even in code size, which makes me believe that a later pass would remove the
duplication (common block elimination on the Cmm level?) and the end
result is not affected.

But in the interest of compilation time it still seems useful to avoid
this duplication from the get-go.